login struturing done
